+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `ikemenExecutable` config to customize the path to Ikemen executable
+- Add `ikemenExecutableWindows` config to customize the path to Ikemen executable on Windows
 
 ### Fixed